Split agent thought endpoint to two
