+++ src/Api/Endpoints/ChannelEndpoints.cs
@@ -1,5 +1,6 @@
 using AzureOpsCrew.Api.Auth;
 using AzureOpsCrew.Api.Background;
+using AzureOpsCrew.Api.Endpoints.Dtos.Agents;
 using AzureOpsCrew.Api.Endpoints.Dtos.Channels;
 using AzureOpsCrew.Api.Endpoints.Dtos.Chats;
 using AzureOpsCrew.Api.Services;
@@ -7,9 +8,11 @@ using AzureOpsCrew.Domain.Agents;
 using AzureOpsCrew.Domain.Channels;
 using AzureOpsCrew.Domain.Chats;
 using AzureOpsCrew.Infrastructure.Ai.Models;
+using AzureOpsCrew.Infrastructure.Ai.Models.Content;
 using AzureOpsCrew.Infrastructure.Db;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
+using System.Text.Json;
 
 namespace AzureOpsCrew.Api.Endpoints;
 
@@ -163,6 +166,123 @@ public static class ChannelEndpoints
             return Results.Ok(messages);
         })
         .Produces<List<Message>>(StatusCodes.Status200OK);
+
+        // GET: /api/channels/{channelId}/agents/{agentId}/mind - Returns agent thoughts scoped to a specific channel
+        group.MapGet("/{channelId}/agents/{agentId}/mind", async (
+            Guid channelId,
+            Guid agentId,
+            AzureOpsCrewContext context,
+            CancellationToken cancellationToken) =>
+        {
+            // Verify channel exists
+            var channel = await context.Channels
+                .SingleOrDefaultAsync(c => c.Id == channelId, cancellationToken);
+
+            if (channel is null)
+                return Results.NotFound();
+
+            // Filter by both AgentId and ThreadId (which equals channelId)
+            var messages = await context.AgentThoughts
+                .Where(m => m.AgentId == agentId && m.ThreadId == channelId && !m.IsHidden)
+                .OrderBy(m => m.CreatedAt)
+                .ToListAsync(cancellationToken);
+
+            var historyMessages = new List<AgentMindEventDto>();
+
+            // First pass: collect tool result content by CallId (from tool-role messages)
+            var toolResultsByCallId = new Dictionary<(Guid threadId, Guid runId, string callId), string>();
+            foreach (var msg in messages)
+            {
+                if (msg.Role.ToString() != "tool")
+                    continue;
+                var contentDto = new AocAiContentDto
+                {
+                    Content = msg.ContentJson,
+                    ContentType = msg.ContentType
+                };
+                var aiContent = contentDto.ToAocAiContent();
+                if (aiContent is AocFunctionResultContent functionResult)
+                {
+                    var resultStr = functionResult.Result switch
+                    {
+                        null => "<null>",
+                        string s => s,
+                        JsonElement el => el.GetRawText(),
+                        _ => JsonSerializer.Serialize(functionResult.Result)
+                    };
+                    toolResultsByCallId[(msg.ThreadId, msg.RunId, functionResult.CallId)] = resultStr ?? "";
+                }
+            }
+
+            foreach (var msg in messages)
+            {
+                // Skip tool role messages (internal function calls)
+                if (msg.Role.ToString() == "tool")
+                    continue;
+
+                // Deserialize content
+                var contentDto = new AocAiContentDto
+                {
+                    Content = msg.ContentJson,
+                    ContentType = msg.ContentType
+                };
+                var aiContent = contentDto.ToAocAiContent();
+
+                if (aiContent is AocTextContent textContent)
+                {
+                    historyMessages.Add(new AgentMindEventDto
+                    {
+                        Id = msg.Id.ToString(),
+                        Role = msg.Role.ToString(),
+                        Content = textContent.Text,
+                        Timestamp = new DateTimeOffset(msg.CreatedAt, TimeSpan.Zero),
+                    });
+                }
+                else if (aiContent is AocTextReasoningContent reasoningContent)
+                {
+                    historyMessages.Add(new AgentMindEventDto
+                    {
+                        Id = msg.Id.ToString(),
+                        Role = msg.Role.ToString(),
+                        Content = null,
+                        Reasoning = reasoningContent.Text,
+                        Timestamp = new DateTimeOffset(msg.CreatedAt, TimeSpan.Zero),
+                    });
+                }
+                else if (aiContent is AocFunctionCallContent functionCallContent
+                         && toolResultsByCallId.TryGetValue((msg.ThreadId, msg.RunId, functionCallContent.CallId), out var resultStr))
+                {
+                    object? resultObj;
+                    try
+                    {
+                        resultObj = JsonSerializer.Deserialize<JsonElement>(resultStr);
+                    }
+                    catch
+                    {
+                        resultObj = new Dictionary<string, object?> { ["raw"] = resultStr };
+                    }
+
+                    historyMessages.Add(new AgentMindEventDto
+                    {
+                        Id = functionCallContent.CallId,
+                        Role = msg.Role.ToString(),
+                        Content = "",
+                        Timestamp = new DateTimeOffset(msg.CreatedAt, TimeSpan.Zero),
+                        Widget = new UiWidgetDto
+                        {
+                            ToolName = functionCallContent.Name,
+                            CallId = functionCallContent.CallId,
+                            Args = functionCallContent.Arguments ?? new Dictionary<string, object?>(),
+                            Result = resultObj
+                        }
+                    });
+                }
+            }
+
+            return Results.Ok(new AgentMindResponseDto { Events = historyMessages });
+        })
+        .Produces<AgentMindResponseDto>(StatusCodes.Status200OK)
+        .Produces(StatusCodes.Status404NotFound);
 
         group.MapPost("/{id}/messages", async (
             Guid id,
